lib: move audio_frame_stereo out of 033_channel_strip
