+++ src/sketches/033_channel_strip/graph/instrument/pan.hpp
@@ -4,7 +4,7 @@
 // Copyright (C) 2025  Rob W. Albus
 // Licensed under the GPLv3. See LICENSE for details.
 
-#include "graph/instrument/audio_frame.hpp"
+#include "lib/audio_frame/audio_frame.hpp"
 
 struct pan_033 {
     audio_frame_stereo pan_coeffs;
